+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Bugfix: Prevent circle/group/gallery names overflowing their visual space
 * Add the ability to un-ban users from circles, groups, and galleries
 * Reduce size of the circle avatar, so "Community" fits on one line on iPad
+* Add ability to rename circles, groups, and galleries
 
 # v1.0.3
 * Bugfix: Ignore invalid config object in account data